Struct crabgrab::capture_stream::CaptureStream

pub struct CaptureStream { /* private fields */ }

Represents an active capture stream

Implementations

impl CaptureStream

pub fn test_access(borderless: bool) -> Option<CaptureAccessToken>

Test whether the calling application has permission to capture content

pub async fn request_access(borderless: bool) -> Option<CaptureAccessToken>

Prompt the user for permission to capture content

pub fn supported_pixel_formats() -> &'static [CapturePixelFormat]

Gets the implementation’s supported pixel formats

pub fn new( token: CaptureAccessToken, config: CaptureConfig, callback: impl FnMut(Result<StreamEvent, StreamError>) + Send + 'static ) -> Result<Self, StreamCreateError>

Start a new capture stream with the given stream callback

pub fn stop(&mut self) -> Result<(), StreamStopError>

Stop the capture