Module crabgrab::feature::metal

Enums

MacosVideoFrameError

Represents an error getting the texture from a video frame

MetalVideoFramePlaneTexture

Identifies planes of a video frame

Traits

MetalCaptureStreamExt

A capture stream which interoperates with Metal

MetalVideoFrameExt

A video frame which can be used to create metal textures