Trait crabgrab::feature::wgpu::WgpuCaptureStreamExt

pub trait WgpuCaptureStreamExt {
    // Required methods
    fn get_wgpu_device_wrapper(
        &self
    ) -> Option<Arc<dyn AsRef<Device> + Send + Sync + 'static>>;
    fn get_wgpu_device(&self) -> Option<&Device>;
}

A capture stream which may have had a Wgpu device instance supplied to it

Required Methods

fn get_wgpu_device_wrapper( &self ) -> Option<Arc<dyn AsRef<Device> + Send + Sync + 'static>>

Gets the Wgpu device wrapper supplied to CaptureConfig::with_wgpu_device(..)

fn get_wgpu_device(&self) -> Option<&Device>

Gets the Wgpu device referenced by device wrapper supplied to CaptureConfig::with_wgpu_device(..)