Trait crabgrab::platform::macos::MacosCapturableWindowExt

pub trait MacosCapturableWindowExt {
    // Required methods
    fn get_window_layer(&self) -> Result<i32, CapturableContentError>;
    fn get_window_level(
        &self
    ) -> Result<MacosWindowLevel, CapturableContentError>;
    fn get_window_id(&self) -> u32;
    fn from_window_id(
        window_id: u32
    ) -> impl Future<Output = Result<CapturableWindow, CapturableContentError>>;
}

Mac OS specific extensions for capturable windows A capturable window with mac-os specific features

Required Methods

fn get_window_layer(&self) -> Result<i32, CapturableContentError>

Get the window layer of this window

fn get_window_level(&self) -> Result<MacosWindowLevel, CapturableContentError>

Get the window level of this window

fn get_window_id(&self) -> u32

Get the native window id for this capturable window. This is the CGWindowID for this window.

fn from_window_id( window_id: u32 ) -> impl Future<Output = Result<CapturableWindow, CapturableContentError>>

Try and convert the given CGWindowID to a capturable window.

Object Safety

This trait is not object safe.